+++ scripts/pre_push_hook_test.py
@@ -23,14 +23,12 @@ import os
 import shutil
 import subprocess
 import sys
-import tempfile
 
-from core import utils
 from core.tests import test_utils
 
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Tuple
 
-from . import common
+from . import git_changes_utils
 from . import install_python_prod_dependencies
 from . import pre_push_hook
 
@@ -52,8 +50,8 @@ class PrePushHookTests(test_utils.GenericTestBase):
             return b'remote'
         def mock_get_refs() -> List[str]:
             return ['ref1', 'ref2']
-        def mock_collect_files_being_pushed(
-            unused_refs: List[pre_push_hook.GitRef],
+        def mock_get_changed_files(
+            unused_refs: List[git_changes_utils.GitRef],
             unused_remote: str
         ) -> Dict[str, Tuple[List[bytes], List[bytes]]]:
             return {
@@ -76,24 +74,18 @@ class PrePushHookTests(test_utils.GenericTestBase):
             return self.mypy_check_code
         self.does_diff_include_js_or_ts_files = False
         def mock_does_diff_include_js_or_ts_files(
-            unused_diff_files: List[pre_push_hook.FileDiff]
+            unused_diff_files: List[git_changes_utils.FileDiff]
         ) -> bool:
             return self.does_diff_include_js_or_ts_files
-        def mock_get_js_or_ts_files_from_diff(
-            unused_diff_files: List[pre_push_hook.FileDiff]
-        ) -> List[str]:
-            if self.does_diff_include_js_or_ts_files:
-                return ['file1.js', 'file2.ts']
-            return []
         self.does_diff_include_ts_files = False
         def mock_does_diff_include_ts_files(
-            unused_diff_files: List[pre_push_hook.FileDiff]
+            unused_diff_files: List[git_changes_utils.FileDiff]
         ) -> bool:
             return self.does_diff_include_ts_files
 
         self.does_diff_include_ci_config_or_js_files = False
         def mock_does_diff_include_ci_config_or_js_files(
-            unused_diff_files: List[pre_push_hook.FileDiff]
+            unused_diff_files: List[git_changes_utils.FileDiff]
         ) -> bool:
             return self.does_diff_include_ci_config_or_js_files
 
@@ -106,11 +98,12 @@ class PrePushHookTests(test_utils.GenericTestBase):
             mock_check_backend_python_library_for_inconsistencies)
         self.popen_swap = self.swap(subprocess, 'Popen', mock_popen)
         self.get_remote_name_swap = self.swap(
-            pre_push_hook, 'get_remote_name', mock_get_remote_name)
-        self.get_refs_swap = self.swap(pre_push_hook, 'get_refs', mock_get_refs)
-        self.collect_files_swap = self.swap(
-            pre_push_hook, 'collect_files_being_pushed',
-            mock_collect_files_being_pushed)
+            git_changes_utils, 'get_remote_name', mock_get_remote_name)
+        self.get_refs_swap = self.swap(
+            git_changes_utils, 'get_refs', mock_get_refs)
+        self.get_changed_files_swap = self.swap(
+            git_changes_utils, 'get_changed_files',
+            mock_get_changed_files)
         self.uncommitted_files_swap = self.swap(
             pre_push_hook, 'has_uncommitted_files', mock_has_uncommitted_files)
         self.print_swap = self.swap(builtins, 'print', mock_print)
@@ -123,9 +116,6 @@ class PrePushHookTests(test_utils.GenericTestBase):
         self.js_or_ts_swap = self.swap(
             pre_push_hook, 'does_diff_include_js_or_ts_files',
             mock_does_diff_include_js_or_ts_files)
-        self.get_js_or_ts_files_swap = self.swap(
-            pre_push_hook, 'get_js_or_ts_files_from_diff',
-            mock_get_js_or_ts_files_from_diff)
         self.ts_swap = self.swap(
             pre_push_hook, 'does_diff_include_ts_files',
             mock_does_diff_include_ts_files)
@@ -139,331 +129,6 @@ class PrePushHookTests(test_utils.GenericTestBase):
             self.assertEqual(
                 pre_push_hook.start_subprocess_for_result(['cmd']),
                 (b'test\n', b''))
-
-    def test_get_remote_name_without_errors(self) -> None:
-        process_for_remote = subprocess.Popen(
-            [b'echo', b'origin\nupstream'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_upstream_url = subprocess.Popen(
-            [b'echo', b'url.oppia/oppia.git'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_origin_url = subprocess.Popen(
-            [b'echo', b'url.other/oppia.git'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        def mock_popen(
-            cmd_tokens: List[bytes], stdout: int, stderr: int  # pylint: disable=unused-argument
-        ) -> subprocess.Popen[bytes]:  # pylint: disable=unsubscriptable-object
-            if b'remote.origin.url' in cmd_tokens:
-                return process_for_origin_url
-            elif b'remote.upstream.url' in cmd_tokens:
-                return process_for_upstream_url
-            else:
-                return process_for_remote
-        popen_swap = self.swap(subprocess, 'Popen', mock_popen)
-        with popen_swap:
-            self.assertEqual(pre_push_hook.get_remote_name(), b'upstream')
-
-    def test_get_remote_name_with_error_in_obtaining_remote(self) -> None:
-        def mock_communicate() -> Tuple[bytes, bytes]:
-            return (b'test', b'test_oppia_error')
-        process = subprocess.Popen(
-            [b'echo', b'test'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        # Here we use MyPy ignore because here we are assigning a value to the
-        # 'communicate' method, and according to MyPy, assignment to a method
-        # is not allowed.
-        process.communicate = mock_communicate  # type: ignore[assignment]
-        def mock_popen(
-            unused_cmd_tokens: List[str], stdout: int, stderr: int  # pylint: disable=unused-argument
-        ) -> subprocess.Popen[bytes]:  # pylint: disable=unsubscriptable-object
-            return process
-
-        popen_swap = self.swap(subprocess, 'Popen', mock_popen)
-        with popen_swap, self.assertRaisesRegex(ValueError, 'test_oppia_error'):
-            pre_push_hook.get_remote_name()
-
-    def test_get_remote_name_with_error_in_obtaining_remote_url(self) -> None:
-        def mock_communicate() -> Tuple[str, str]:
-            return ('test', 'test_oppia_error')
-        process_for_remote = subprocess.Popen(
-            [b'echo', b'origin\nupstream'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_remote_url = subprocess.Popen(
-            [b'echo', b'test'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        def mock_popen(
-            cmd_tokens: List[bytes], stdout: int, stderr: int  # pylint: disable=unused-argument
-        ) -> subprocess.Popen[bytes]:  # pylint: disable=unsubscriptable-object
-            if b'config' in cmd_tokens:
-                return process_for_remote_url
-            else:
-                return process_for_remote
-
-        communicate_swap = self.swap(
-            process_for_remote_url, 'communicate', mock_communicate
-        )
-        popen_swap = self.swap(subprocess, 'Popen', mock_popen)
-        with communicate_swap:
-            with popen_swap:
-                with self.assertRaisesRegex(ValueError, 'test_oppia_error'):
-                    pre_push_hook.get_remote_name()
-
-    def test_get_remote_name_with_no_remote_set(self) -> None:
-        process_for_remote = subprocess.Popen(
-            [b'echo', b'origin\nupstream'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_upstream_url = subprocess.Popen(
-            [b'echo', b'url.other/oppia.git'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_origin_url = subprocess.Popen(
-            [b'echo', b'url.other/oppia.git'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        def mock_popen(
-            cmd_tokens: List[bytes], stdout: int, stderr: int  # pylint: disable=unused-argument
-        ) -> subprocess.Popen[bytes]:  # pylint: disable=unsubscriptable-object
-            if b'remote.origin.url' in cmd_tokens:
-                return process_for_origin_url
-            elif b'remote.upstream.url' in cmd_tokens:
-                return process_for_upstream_url
-            else:
-                return process_for_remote
-        popen_swap = self.swap(subprocess, 'Popen', mock_popen)
-        with popen_swap, self.assertRaisesRegex(
-            Exception,
-            'Error: Please set upstream for the lint checks to run '
-            'efficiently. To do that follow these steps:\n'
-            '1. Run the command \'git remote -v\'\n'
-            '2a. If upstream is listed in the command output, then run the '
-            'command \'git remote set-url upstream '
-            'https://github.com/oppia/oppia.git\'\n'
-            '2b. If upstream is not listed in the command output, then run the '
-            'command \'git remote add upstream '
-            'https://github.com/oppia/oppia.git\'\n'):
-            pre_push_hook.get_remote_name()
-
-    def test_get_remote_name_with_multiple_remotes_set(self) -> None:
-        process_for_remote = subprocess.Popen(
-            [b'echo', b'origin\nupstream'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_upstream_url = subprocess.Popen(
-            [b'echo', b'url.oppia/oppia.git'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        process_for_origin_url = subprocess.Popen(
-            [b'echo', b'url.oppia/oppia.git'], stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE)
-        def mock_popen(
-            cmd_tokens: List[bytes], stdout: int, stderr: int  # pylint: disable=unused-argument
-        ) -> subprocess.Popen[bytes]:  # pylint: disable=unsubscriptable-object
-            if b'remote.origin.url' in cmd_tokens:
-                return process_for_origin_url
-            elif b'remote.upstream.url' in cmd_tokens:
-                return process_for_upstream_url
-            else:
-                return process_for_remote
-        popen_swap = self.swap(subprocess, 'Popen', mock_popen)
-        with popen_swap, self.print_swap:
-            self.assertIsNone(pre_push_hook.get_remote_name())
-        self.assertTrue(
-            'Warning: Please keep only one remote branch for oppia:develop '
-            'to run the lint checks efficiently.\n' in self.print_arr)
-
-    def test_git_diff_name_status_without_error(self) -> None:
-        def mock_start_subprocess_for_result(
-            unused_cmd_tokens: List[str]
-        ) -> Tuple[bytes, None]:
-            return (b'M\tfile1\nA\tfile2', None)
-        subprocess_swap = self.swap(
-            pre_push_hook, 'start_subprocess_for_result',
-            mock_start_subprocess_for_result)
-
-        with subprocess_swap:
-            self.assertEqual(
-                pre_push_hook.git_diff_name_status(
-                    'left', 'right', diff_filter='filter'),
-                [
-                    pre_push_hook.FileDiff(status=b'M', name=b'file1'),
-                    pre_push_hook.FileDiff(status=b'A', name=b'file2')])
-
-    def test_git_diff_name_status_with_error(self) -> None:
-        def mock_start_subprocess_for_result(
-            unused_cmd_tokens: List[str]
-        ) -> Tuple[str, str]:
-            return ('M\tfile1\nA\tfile2', 'test_oppia_error')
-        subprocess_swap = self.swap(
-            pre_push_hook, 'start_subprocess_for_result',
-            mock_start_subprocess_for_result)
-
-        with subprocess_swap, self.assertRaisesRegex(
-            ValueError, 'test_oppia_error'
-        ):
-            pre_push_hook.git_diff_name_status(
-                'left', 'right', diff_filter='filter')
-
-    def test_compare_to_remote(self) -> None:
-        check_function_calls = {
-            'start_subprocess_for_result_is_called': False,
-            'git_diff_name_status_is_called': False,
-            'get_merge_base_is_called': False,
-        }
-        expected_check_function_calls = {
-            'start_subprocess_for_result_is_called': True,
-            'git_diff_name_status_is_called': True,
-            'get_merge_base_is_called': True,
-        }
-        def mock_start_subprocess_for_result(
-            unused_cmd_tokens: List[str]
-        ) -> None:
-            check_function_calls['start_subprocess_for_result_is_called'] = True
-        def mock_git_diff_name_status(
-            unused_left: str, unused_right: str
-        ) -> str:
-            check_function_calls['git_diff_name_status_is_called'] = True
-            return 'Test'
-        def mock_get_merge_base(unused_left: str, unused_right: str) -> str:
-            check_function_calls['get_merge_base_is_called'] = True
-            return 'Merge Base'
-        subprocess_swap = self.swap(
-            pre_push_hook, 'start_subprocess_for_result',
-            mock_start_subprocess_for_result)
-        git_diff_swap = self.swap(
-            pre_push_hook, 'git_diff_name_status', mock_git_diff_name_status)
-        get_merge_base_swap = self.swap(
-            pre_push_hook, 'get_merge_base', mock_get_merge_base)
-
-        with subprocess_swap, git_diff_swap, get_merge_base_swap:
-            self.assertEqual(
-                pre_push_hook.compare_to_remote('remote', 'local branch'),
-                'Test'
-            )
-        self.assertEqual(check_function_calls, expected_check_function_calls)
-
-    def test_get_merge_base_reports_error(self) -> None:
-        def mock_start_subprocess_for_result(
-            unused_cmd_tokens: List[str]
-        ) -> Tuple[None, str]:
-            return None, 'Test'
-        subprocess_swap = self.swap(
-            pre_push_hook, 'start_subprocess_for_result',
-            mock_start_subprocess_for_result)
-
-        with subprocess_swap, self.assertRaisesRegex(ValueError, 'Test'):
-            pre_push_hook.get_merge_base('A', 'B')
-
-    def test_get_merge_base_returns_merge_base(self) -> None:
-        check_function_calls = {
-            'start_subprocess_for_result_is_called': False,
-        }
-        expected_check_function_calls = {
-            'start_subprocess_for_result_is_called': True,
-        }
-        def mock_start_subprocess_for_result(
-            unused_cmd_tokens: List[str]
-        ) -> Tuple[bytes, None]:
-            check_function_calls['start_subprocess_for_result_is_called'] = True
-            return b'Test', None
-        subprocess_swap = self.swap(
-            pre_push_hook, 'start_subprocess_for_result',
-            mock_start_subprocess_for_result)
-
-        with subprocess_swap:
-            self.assertEqual(pre_push_hook.get_merge_base('A', 'B'), 'Test')
-
-        self.assertEqual(check_function_calls, expected_check_function_calls)
-
-    def test_extract_files_to_lint_with_empty_file_diffs(self) -> None:
-        self.assertEqual(pre_push_hook.extract_files_to_lint([]), [])
-
-    def test_extract_files_to_lint_with_non_empty_file_diffs(self) -> None:
-        self.assertEqual(
-            pre_push_hook.extract_files_to_lint([
-                pre_push_hook.FileDiff(status=b'M', name=b'file1'),
-                pre_push_hook.FileDiff(status=b'A', name=b'file2'),
-                pre_push_hook.FileDiff(status=b'W', name=b'file3')]),
-            [b'file1', b'file2'])
-
-    def test_get_parent_branch_name_for_diff_with_hotfix_branch(self) -> None:
-        def mock_get_branch() -> str:
-            return 'release-1.2.3-hotfix-1'
-        get_branch_swap = self.swap(
-            common, 'get_current_branch_name', mock_get_branch)
-        with get_branch_swap:
-            self.assertEqual(
-                pre_push_hook.get_parent_branch_name_for_diff(),
-                'release-1.2.3')
-
-    def test_get_parent_branch_name_for_diff_with_release_branch(self) -> None:
-        def mock_get_branch() -> str:
-            return 'release-1.2.3'
-        get_branch_swap = self.swap(
-            common, 'get_current_branch_name', mock_get_branch)
-        with get_branch_swap:
-            self.assertEqual(
-                pre_push_hook.get_parent_branch_name_for_diff(), 'develop')
-
-    def test_get_parent_branch_name_for_diff_with_non_release_branch(
-        self
-    ) -> None:
-        def mock_get_branch() -> str:
-            return 'branch-1'
-        get_branch_swap = self.swap(
-            common, 'get_current_branch_name', mock_get_branch)
-        with get_branch_swap:
-            self.assertEqual(
-                pre_push_hook.get_parent_branch_name_for_diff(), 'develop')
-
-    def test_collect_files_being_pushed_with_empty_ref_list(self) -> None:
-        def mock_get_branch() -> str:
-            return 'branch-1'
-        get_branch_swap = self.swap(
-            common, 'get_current_branch_name', mock_get_branch)
-        with get_branch_swap:
-            self.assertEqual(
-                pre_push_hook.collect_files_being_pushed([], 'remote'), {})
-
-    def test_collect_files_being_pushed_with_non_empty_ref_list(self) -> None:
-        def mock_get_branch() -> str:
-            return 'branch-1'
-        def mock_compare_to_remote(
-            unused_remote: str,
-            unused_local_branch: str,
-            remote_branch: Optional[str] = None  # pylint: disable=unused-argument
-        ) -> List[str]:
-            return ['A:file1', 'M:file2']
-        def mock_extract_files_to_lint(
-            unused_file_diffs: List[pre_push_hook.FileDiff]
-        ) -> List[str]:
-            return ['file1', 'file2']
-
-        get_branch_swap = self.swap(
-            common, 'get_current_branch_name', mock_get_branch)
-        compare_to_remote_swap = self.swap(
-            pre_push_hook, 'compare_to_remote', mock_compare_to_remote)
-        extract_files_swap = self.swap(
-            pre_push_hook, 'extract_files_to_lint', mock_extract_files_to_lint)
-
-        with compare_to_remote_swap, extract_files_swap, get_branch_swap:
-            self.assertEqual(
-                pre_push_hook.collect_files_being_pushed([
-                    pre_push_hook.GitRef(
-                        local_ref='refs/heads/branch1', local_sha1='sha1',
-                        remote_ref='remote/ref1', remote_sha1='rsha1'),
-                    pre_push_hook.GitRef(
-                        local_ref='refs/branch2', local_sha1='sha2',
-                        remote_ref='remote/ref2', remote_sha1='rsha2')
-                    ], 'remote'),
-                {'branch1': (['A:file1', 'M:file2'], ['file1', 'file2'])})
-
-    def test_get_refs(self) -> None:
-        temp_stdin_file = tempfile.NamedTemporaryFile().name
-        with utils.open_file(temp_stdin_file, 'w') as f:
-            f.write('local_ref local_sha1 remote_ref remote_sha1')
-        with utils.open_file(temp_stdin_file, 'r') as f:
-            with self.swap(sys, 'stdin', f):
-                self.assertEqual(
-                    pre_push_hook.get_refs(),
-                    [
-                        pre_push_hook.GitRef(
-                            local_ref='local_ref', local_sha1='local_sha1',
-                            remote_ref='remote_ref', remote_sha1='remote_sha1'
-                        )])
 
     def test_start_linter(self) -> None:
         with self.popen_swap:
@@ -653,18 +318,6 @@ class PrePushHookTests(test_utils.GenericTestBase):
             pre_push_hook.does_diff_include_js_or_ts_files(
                 [b'file1.html', b'file2.py']))
 
-    def test_get_js_or_ts_files_from_diff_with_js_file(self) -> None:
-        self.assertEqual(
-            pre_push_hook.get_js_or_ts_files_from_diff(
-                [b'file1.js', b'file2.ts', b'file3.py']),
-            ['file1.js', 'file2.ts'])
-
-    def test_get_js_or_ts_files_from_diff_with_no_file(self) -> None:
-        self.assertEqual(
-            pre_push_hook.get_js_or_ts_files_from_diff(
-                [b'file1.html', b'file2.py', b'file3.py']),
-            [])
-
     def test_does_diff_include_ts_files(self) -> None:
         self.assertTrue(
             pre_push_hook.does_diff_include_ts_files(
@@ -692,7 +345,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
         uncommitted_files_swap = self.swap(
             pre_push_hook, 'has_uncommitted_files', mock_has_uncommitted_files)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, uncommitted_files_swap:
+            with self.get_changed_files_swap, uncommitted_files_swap:
                 with self.assertRaisesRegex(SystemExit, '1'):
                     with self.swap_check_backend_python_libs:
                         pre_push_hook.main(args=[])
@@ -711,7 +364,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
         check_output_swap = self.swap(
             subprocess, 'check_output', mock_check_output)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with check_output_swap, self.assertRaisesRegex(
                     SystemExit, '1'
                 ):
@@ -727,7 +380,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
     def test_lint_failure(self) -> None:
         self.linter_code = 1
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with self.execute_mypy_checks_swap:
                         with self.assertRaisesRegex(SystemExit, '1'):
@@ -740,7 +393,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
     def test_mypy_check_failure(self) -> None:
         self.mypy_check_code = 1
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with self.execute_mypy_checks_swap:
                         with self.assertRaisesRegex(SystemExit, '1'):
@@ -760,7 +413,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
             pre_push_hook, 'run_script_and_get_returncode',
             mock_run_script_and_get_returncode)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with self.ts_swap, run_script_and_get_returncode_swap:
                         with self.execute_mypy_checks_swap:
@@ -781,7 +434,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
             pre_push_hook, 'run_script_and_get_returncode',
             mock_run_script_and_get_returncode)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with self.ts_swap, run_script_and_get_returncode_swap:
                         with self.execute_mypy_checks_swap:
@@ -802,7 +455,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
             mock_run_script_and_get_returncode)
 
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with self.ts_swap, run_script_and_get_returncode_swap:
                         with self.execute_mypy_checks_swap:
@@ -816,24 +469,20 @@ class PrePushHookTests(test_utils.GenericTestBase):
     def test_frontend_test_failure(self) -> None:
         self.does_diff_include_js_or_ts_files = True
         def mock_run_script_and_get_returncode(script: List[str]) -> int:
-            frontend_test_cmds = pre_push_hook.FRONTEND_TEST_CMDS.copy()
-            frontend_test_cmds.append(
-                '--specs_to_run=file1.js,file2.ts')
-            if script == frontend_test_cmds:
+            if script == pre_push_hook.FRONTEND_TEST_CMDS:
                 return 1
             return 0
         run_script_and_get_returncode_swap = self.swap(
             pre_push_hook, 'run_script_and_get_returncode',
             mock_run_script_and_get_returncode)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
-                with self.check_output_swap, self.get_js_or_ts_files_swap:
-                    with self.js_or_ts_swap, self.start_linter_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
+                with self.check_output_swap, self.start_linter_swap:
+                    with self.js_or_ts_swap, run_script_and_get_returncode_swap:
                         with self.execute_mypy_checks_swap:
-                            with run_script_and_get_returncode_swap:
-                                with self.assertRaisesRegex(SystemExit, '1'):
-                                    with self.swap_check_backend_python_libs:
-                                        pre_push_hook.main(args=[])
+                            with self.assertRaisesRegex(SystemExit, '1'):
+                                with self.swap_check_backend_python_libs:
+                                    pre_push_hook.main(args=[])
         self.assertTrue(
             'Push aborted due to failing frontend tests.' in self.print_arr)
 
@@ -848,7 +497,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
             pre_push_hook, 'run_script_and_get_returncode',
             mock_run_script_and_get_returncode)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with run_script_and_get_returncode_swap:
                         with self.ci_config_or_js_files_swap:
@@ -878,7 +527,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
             pre_push_hook, 'run_script_and_get_returncode',
             mock_run_script_and_get_returncode)
         with self.get_remote_name_swap, self.get_refs_swap, self.print_swap:
-            with self.collect_files_swap, self.uncommitted_files_swap:
+            with self.get_changed_files_swap, self.uncommitted_files_swap:
                 with self.check_output_swap, self.start_linter_swap:
                     with run_script_and_get_returncode_swap:
                         with self.js_or_ts_swap:
